mcp/pyproject.toml
[project]
name = "logistics-mcp-server"
version = "0.1.0"
description = "MCP server for logistics flight data (HTTP transport)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.25.0",
    "httpx>=0.27.0,<1.0",
    "uvicorn>=0.27.0",
    "starlette>=0.40.0",
    "python-dotenv",
    "duckdb>=1.0.0",
    "pyjwt[crypto]>=2.8.0",
    "cachetools>=5.3.0",
]

[tool.uv]
prerelease = "allow"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
